+++ oshi-common/src/main/java/oshi/hardware/common/platform/linux/LinuxHWDiskStore.java
@@ -83,6 +83,10 @@ public abstract class LinuxHWDiskStore extends AbstractHWDiskStore {
         super(name, model, serial, size);
     }
 
+    protected LinuxHWDiskStore(String name, String model, String serial, long size, String diskType) {
+        super(name, model, serial, size, diskType);
+    }
+
     @Override
     public long getReads() {
         return reads;